NFC: simplify error handling with arena allocator
This is conceptually a non-functional change (NFC), but practically
speaking, there is significant potential for regressions because of how
many codepaths this touches.

Prior to this change, each individual function in src/passes/*.c was
reponsible for cleaning up any heap-allocated data structures whose
ownership was not transferred up the callstack. In a bit of irony, the
common way for this kind of garbage to be created was error-handling for
allocation failures themselves, like:

    struct example *left = malloc(...);
    assert(left != NULL);

    struct example *right = malloc(...);
    if (right == NULL) {
        return foobar; /* leaks <left> */
    }

    return recurse(left, right);

Use of __attribute__((cleanup(...))) seemed to work well in the past,
but as the complexity of src/passes/*.c has grown, clang-analyzer seems
to be getting confused more often. This has required more frequent use
of NOLINT(clang-analyzer-unix.Malloc) comments, in some cases even on
codepaths that are not themselves changing (but which are presumably
getting compiled or optimized differently depending on other code).

Overall, we can avoid these recurring, spurious clang-analyzer warnings
and make future errors on rarely-exercised error codepaths less likely
by simplifying the ownership model. With an arena-oriented design, each
pass can avoid dealing with the details of "orphaned" AST nodes, IR ops,
etc. caused by exceptional situations and instead rely on the API
consumer to arena_free() when appropriate, which handles both live and
"dead" data (not truly leaked memory, just not reachable from the output
AST, IR, etc. datastructures).